+++ src/main/java/com/movietix/xiazihao/entity/pojo/Order.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,7 +14,7 @@ public class Order {
     private String orderNo; // 订单编号
     private Integer userId; // 用户ID
     private Integer screeningId; // 场次ID
-    private BigDecimal totalAmount = BigDecimal.valueOf(0); // 总金额
+    private Double totalAmount = 0.0; // 总金额
     private Integer seatCount = 0; // 座位数
     private Integer status = 0; // 状态:0-待支付 1-已支付 2-已取消 3-已完成 4-已退款
     private String contactPhone; // 联系电话
